レベルシステムをRepositoryに移行
